move dataset fixtures to conftest
